feat(database): Change book series position to Double

+++ feature/books/data/src/commonMain/kotlin/com/retro99/books/data/model/SeriesApiModel.kt
@@ -16,7 +16,7 @@ data class SeriesApiModel(
     val featured: Int? = null,
 
     @SerialName("position")
-    val position: Int? = null,
+    val position: Double? = null,
 
     @SerialName("createdAt")
     val createdAt: String? = null,
