+++ feature/books/data/src/commonMain/kotlin/com/retro99/books/data/model/SeriesWithPositionLocalModel.kt
@@ -7,7 +7,7 @@ data class SeriesWithPositionLocalModel(
     override val uuid: String,
     override val name: String,
     override val featured: Int?,
-    override val position: Int?,
+    override val position: Double?,
     override val createdAt: String?,
     override val updatedAt: String?,
 ) : SeriesWithPositionEntity
